PooJava/src/Objetos/AulaTest.java
package Objetos;

import org.junit.Assert;
import org.junit.Test;

import java.util.ArrayList;
import java.util.Date;
import java.util.List;

public class AulaTest {

    @Test
    public void deveCriarAulaPorMeiodoConstrutor() {

        Aluno aluno1 = new Aluno("tercio25@gmail.com", "Tercio Rodrigues");
        Aluno aluno2 = new Aluno( "joaosilva@gmail.com", "João Silva");
        Professor professor = new Professor ( "igorsantos@gmail.com", "Igor Santos");
        Materia materia = new Materia(
                "Materia Test",
                "Descrição Materia",
                10.5,
                professor,
                3);
        Date data = new Date();
        List<Aluno> alunos = new ArrayList<>();
        alunos.add(aluno1);
        alunos.add(aluno2);
        Aula aula = new Aula(data, materia, alunos);

        Assert.assertEquals(data, aula.getData());
        Assert.assertEquals(materia, aula.getMateria());
        Assert.assertEquals(alunos, aula.getListaAlunos());
    }

}

